Handle timestamp offsets
- This required a bit of massaging of type signatures, since date
  granularity is needed in `Info` parsing, even though the granularity
  value actually appears *after* any `PrimitiveBlock`.

+++ src/Streaming/Osm/Types.hs
@@ -69,7 +69,7 @@ memtype 2 = R
 -- many of these fields may be missing for older Elements.
 data Info = Info { _id        :: Int
                  , _version   :: Int
-                 , _timestamp :: Maybe Int
+                 , _timestamp :: Maybe Int64
                  , _changeset :: Maybe Int
                  , _uid       :: Maybe Int
                  , _username  :: Maybe B.ByteString
